+++ packages/firebase/index.d.ts
@@ -6920,6 +6920,19 @@ declare namespace firebase.database.ServerValue {
    * ```
    */
   var TIMESTAMP: Object;
+  
+  /**
+   * Returns a placeholder value that can be used to atomically increment the 
+   * current database value by the provided delta.
+   *
+   * @param delta the amount to modify the current value atomically.
+   * @return a placeholder value for modifying data atomically server-side.
+   */
+  function increment(delta: number) : {
+    '.sv': {
+      'increment': number
+    }
+  };
 }
 
 /**