+++ packages/database/CHANGELOG.md
@@ -1,7 +1,10 @@
 # Unreleased
+- [feature] Added ServerValue.increment() to support atomic field value increments
+  without transactions.
+
+# Released
 - [fixed] Fixed an issue that caused large numeric values with leading zeros to
   not always be sorted correctly.
-
 - [changed] Internal cleanup to Node.JS support.
 
 # 6.4.0